move rei-compat to base so @Luligabi1 would suffer less
