+++ base/build.gradle.kts
@@ -1,4 +1,5 @@
 val libguiVersion: String by project
+val reiVersion: String by project
 //val patchouliVersion: String by project
 
 repositories {
@@ -10,11 +11,21 @@ repositories {
         name = "Patchouli"
         url = uri("https://maven.blamejared.com")
     }*/
+    maven {
+        name = "Shedaniel's Maven Repository"
+        url = uri("https://maven.shedaniel.me/")
+    }
+    maven {
+        name = "TerraformersMC"
+        url = uri("https://maven.terraformersmc.com/releases/")
+    }
 }
 
 dependencies {
     modApi("io.github.cottonmc:LibGui:${libguiVersion}")
     include("io.github.cottonmc:LibGui:${libguiVersion}")
+
+    modApi("me.shedaniel:RoughlyEnoughItems-fabric:${reiVersion}")
 
     //modApi("vazkii.patchouli:Patchouli:${patchouliVersion}")
 }
