+++ base/src/main/kotlin/dev/nathanpb/dml/compat/rei/display/CrushingRecipeDisplay.kt
