+++ base/src/main/resources/fabric.mod.json
@@ -14,7 +14,8 @@
   "environment": "*",
   "entrypoints": {
     "main": ["dev.nathanpb.dml.DeepMobLearningKt::init"],
-    "client": ["dev.nathanpb.dml.DeepMobLearningKt::initClient"]
+    "client": ["dev.nathanpb.dml.DeepMobLearningKt::initClient"],
+    "rei_client": ["dev.nathanpb.dml.compat.rei.ReiPlugin"]
   },
   "mixins": ["mixins.dml-refabricated-base.json"],
   "depends": {
