--- rei-compat/src/main/java/dev/nathanpb/dml/compat/rei/mixin/LivingEntityMixin.java
@@ -1,53 +1,0 @@
-package dev.nathanpb.dml.compat.rei.mixin;
-
-/*
- * Copyright (C) 2020 Nathan P. Bombana, IterationFunk
- *
- * This file is part of Deep Mob Learning: Refabricated.
- *
- * Deep Mob Learning: Refabricated is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- * Deep Mob Learning: Refabricated is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with Deep Mob Learning: Refabricated.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-import dev.nathanpb.dml.compat.rei.accessor.ILivingEntityReiStateAccessor;
-import net.minecraft.entity.EquipmentSlot;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.item.ItemStack;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(LivingEntity.class)
-public class LivingEntityMixin implements ILivingEntityReiStateAccessor {
-    boolean dmlRefIsInREIScreen = false;
-
-    @Override
-    public boolean isDmlRefIsInReiScreen() {
-        return dmlRefIsInREIScreen;
-    }
-
-    @Override
-    public void setDmlRefInReiScreen(boolean flag) {
-        dmlRefIsInREIScreen = flag;
-    }
-
-
-    @Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/LivingEntity;playEquipSound(Lnet/minecraft/item/ItemStack;)V"), method = "onEquipStack", cancellable = true)
-    public void onEquip(EquipmentSlot slot, ItemStack oldStack, ItemStack newStack, CallbackInfo ci) {
-        if (isDmlRefIsInReiScreen()) {
-            ci.cancel();
-        }
-    }
-
-}
