--- rei-compat/src/main/resources/fabric.mod.json
@@ -1,30 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "dml-refabricated-rei",
-  "version": "${version}",
-  "name": "DML: Refabricated - REI Integration",
-  "description": "REI integration plugin for Deep Mob Learning: Refabricated's recipes & Trial system.",
-  "authors": ["NathanPB", "Luligabi"],
-  "contact": {
-    "homepage": "https://www.curseforge.com/minecraft/mc-mods/deep-mob-learning-refabricated",
-    "repo": "https://github.com/CafeteriaGuild/DeepMobLearning-Refabricated",
-    "issues": "https://github.com/CafeteriaGuild/DeepMobLearning-Refabricated/issues"
-  },
-  "license": "LGPL-v3",
-  "environment": "*",
-  "entrypoints": {
-    "rei_client": ["dev.nathanpb.dml.compat.rei.ReiPlugin"]
-  },
-  "mixins": ["mixins.dml-refabricated-rei.json"],
-  "depends": {
-    "dml-refabricated": "${version}"
-  },
-  "suggests": {
-    "roughlyenoughitems": ">=6.0.262-alpha"
-  },
-  "custom": {
-    "modmenu": {
-      "parent": "dml-refabricated"
-    }
-  }
-}
