--- rei-compat/src/main/resources/mixins.dml-refabricated-rei.json
@@ -1,12 +1,0 @@
-{
-  "required": true,
-  "package": "dev.nathanpb.dml.compat.rei.mixin",
-  "compatibilityLevel": "JAVA_8",
-  "mixins": [
-    "LivingEntityMixin"
-  ],
-  "client": [],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
